refactor: move all internal functions to functions.go
